Fixed two problems of schedulling
- Multiple activations, closes #10
- Check Registry not relying on a command by itself but forcing input, closes #8
